[refactor] tui.py として別ファイルにした
