+++ src/redi/cli.py
@@ -65,6 +65,7 @@ from redi.wiki import (
     read_wiki,
     update_wiki,
 )
+from redi.tui import run_issue_tui
 
 
 def build_wiki_tree_choices(pages: list[dict]) -> list[questionary.Choice]:
@@ -1015,104 +1016,6 @@ def _handle_time_entry(args: argparse.Namespace) -> None:
         list_time_entries(project_id=project_id, user_id=args.user_id, full=args.full)
 
 
-def _handle_tui() -> None:
-    from prompt_toolkit import Application
-    from prompt_toolkit.key_binding import KeyBindings
-    from prompt_toolkit.layout import Layout
-    from prompt_toolkit.layout.containers import HSplit, Window
-    from prompt_toolkit.layout.controls import FormattedTextControl
-
-    page_size = 20
-    offset = 0
-    cursor = 0
-    issues = fetch_issues(project_id=default_project_id, limit=page_size, offset=offset)
-    if not issues:
-        print("イシューが見つかりません")
-        return
-
-    def render_issues():
-        result = []
-        for i, issue in enumerate(issues):
-            text = f"#{issue['id']} {issue['subject']}\n"
-            result.append(("reverse" if i == cursor else "", text))
-        return result
-
-    def render_status():
-        page = offset // page_size + 1
-        return [
-            (
-                "reverse",
-                f" Page {page} (offset={offset})  "
-                "↑↓/jk:移動 ←→/hl:ページ Enter:表示 q:終了 ",
-            )
-        ]
-
-    kb = KeyBindings()
-
-    @kb.add("up")
-    @kb.add("k")
-    def _(event):
-        nonlocal cursor
-        cursor = max(0, cursor - 1)
-
-    @kb.add("down")
-    @kb.add("j")
-    def _(event):
-        nonlocal cursor
-        cursor = min(len(issues) - 1, cursor + 1)
-
-    @kb.add("right")
-    @kb.add("l")
-    def _(event):
-        nonlocal offset, cursor, issues
-        next_issues = fetch_issues(
-            project_id=default_project_id,
-            limit=page_size,
-            offset=offset + page_size,
-        )
-        if next_issues:
-            offset += page_size
-            issues = next_issues
-            cursor = 0
-
-    @kb.add("left")
-    @kb.add("h")
-    def _(event):
-        nonlocal offset, cursor, issues
-        if offset > 0:
-            offset = max(0, offset - page_size)
-            issues = fetch_issues(
-                project_id=default_project_id, limit=page_size, offset=offset
-            )
-            cursor = 0
-
-    @kb.add("enter")
-    def _(event):
-        event.app.exit(result=str(issues[cursor]["id"]))
-
-    @kb.add("q")
-    @kb.add("escape")
-    @kb.add("c-c")
-    def _(event):
-        event.app.exit(result=None)
-
-    app = Application(
-        layout=Layout(
-            HSplit(
-                [
-                    Window(FormattedTextControl(render_issues)),
-                    Window(FormattedTextControl(render_status), height=1),
-                ]
-            )
-        ),
-        key_bindings=kb,
-        full_screen=True,
-    )
-    result = app.run()
-    if result is not None:
-        read_issue(result)
-
-
 def main() -> None:
     parser, a_parser = _build_parser()
     argcomplete.autocomplete(parser)
@@ -1125,7 +1028,7 @@ def main() -> None:
     check_config()
 
     if args.tui and args.command is None:
-        _handle_tui()
+        run_issue_tui()
         return
 
     if args.command in ("project", "p"):
